:truck: Move types to shared folder
