+++ src/Changelogs/ChangelogDatastore.ts
@@ -2,7 +2,7 @@ import { Db } from 'mongodb';
 import { COLLECTIONS } from '../drivers/MongoDriver';
 import { reportError } from '../drivers/SentryConnector';
 import { ResourceError, ResourceErrorReason, ServiceError, ServiceErrorReason } from '../errors';
-import { ChangeLogDocument } from '../types/Changelog';
+import { ChangeLogDocument } from '../shared/types/Changelog';
 
 export class ChangelogDataStore {
   constructor(private db: Db) { }
