+++ src/Changelogs/ChangelogRouteDriver.ts
@@ -1,7 +1,7 @@
 import { Request, Response, Router } from 'express';
 import { mapErrorToResponseData } from '../errors';
 import { DataStore } from '../shared/interfaces/DataStore';
-import { UserToken } from '../types';
+import { UserToken } from '../shared/types';
 import * as ChangelogInteractor from './ChangelogInteractor';
 
 /**
