+++ src/LearningObjectSubmission/SubmissionDatastore.ts
@@ -1,7 +1,7 @@
 import { COLLECTIONS } from '../drivers/MongoDriver';
 import { Db } from 'mongodb';
 import * as ObjectMapper from '../drivers/Mongo/ObjectMapper';
-import { UserDocument } from '../types';
+import { UserDocument } from '../shared/types';
 import { LearningObject, User } from '../entity';
 import { Submission } from './types/Submission';
 import { ResourceError, ResourceErrorReason } from '../errors';
