+++ src/LearningObjects/AuthorizationManager.ts
@@ -3,7 +3,7 @@
  */
 
 import { ResourceError, ResourceErrorReason } from '../errors';
-import { UserToken } from '../types';
+import { UserToken } from '../shared/types';
 
 enum AccessGroup {
   ADMIN = 'admin',
