+++ src/shared/types/learning-outcome-document.ts
