+++ src/shared/types/service-token.ts
