+++ src/shared/types/standard-outcome-document.ts
